Remove cached disk ids
This was premature optimization on my part. Was not working when you
mounted a second disk in FN-PC after booting with 1 disk, as the cache was
not getting cleared (and could not be easily). As the operation is fairly light weight
to request which devices are disks, it does not add too much overhead